Abstract CssCodeEditor
