+++ src/components/custom-css-control/editor.scss
@@ -19,52 +19,4 @@
 			margin-top: 7px;
 		}
 	}
-	.maxi-base-control__field {
-		flex-direction: row;
-		flex-wrap: wrap;
-
-		.components-button {
-			margin-left: 10px;
-			float: right;
-			height: 24px;
-			background-color: var(--maxi-grey-9-color);
-			border: 1px solid var(--maxi-grey-1-color);
-			border: 1px solid var(--maxi-grey-1-color);
-			border-radius: 0;
-		}
-	}
-	.w-tc-editor {
-		margin: 8px;
-		width: 100%;
-		background-color: var(--maxi-white-4-color);
-		border: 1px solid var(--maxi-grey-1-color);
-		font-family: monospace !important;
-		flex: 1 0 100%;
-
-		textarea {
-			background-color: var(--maxi-white-4-color);
-			font-family: monospace !important;
-		}
-	}
-	code {
-		padding: 0;
-		background-color: var(--maxi-white-4-color);
-		font-family: monospace !important;
-		span {
-			background-color: var(--maxi-white-4-color);
-			font-family: monospace !important;
-		}
-	}
-	.maxi-custom-css-control__error {
-		margin-top: 8px;
-		padding-left: 10px;
-		padding-right: 5px;
-		font-style: italic;
-		&.not-valid {
-			border-left: 2px solid var(--maxi-primary-color);
-		}
-		&.valid {
-			border-left: 2px solid var(--maxi-green-1-color);
-		}
-	}
 }
