+++ src/components/index.js
@@ -23,6 +23,7 @@ export { default as ClipPathControl } from './clip-path-control';
 export { default as ColorControl } from './color-control';
 export { default as ContentLoader } from './content-loader';
 export { default as ContextLoop } from './context-loop';
+export { default as CssCodeEditor } from './css-code-editor';
 export { default as CustomCss } from './custom-css-control';
 export { default as CustomLabel } from './custom-label';
 export { default as DefaultStylesControl } from './default-styles-control';
